refactor: move window tests to dedicated tests.rs module
Extract the #[cfg(test)] mod tests block (~3500 lines) from
window/mod.rs into window/tests.rs. This reduces mod.rs from
6244 to 2733 lines while keeping all 69 tests (3 active + 66
ignored) in their original module path.

Fixes #99
